use int instead Int64

+++ example/lib/whoopwhoop/ZoomMacOSImplementation.dart
@@ -1,5 +1,4 @@
 import 'dart:async';
-import 'package:fixnum/fixnum.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'ZoomPlatformInterface.dart';
@@ -26,7 +25,7 @@ class ZoomMacOSImplementation implements ZoomPlatformInterface {
 
   @override
   Future<bool?> joinMeeting({
-    required Int64 meetingNumber,
+    required int meetingNumber,
     required String meetingPassword,
     required String displayName
   }) async {
